requirements.txt
fastapi
uvicorn
wikipedia
python-dotenv
google-generativeai
